TOC related code is factored out
